// src/app/generated_services/model/carlonGracieBackendFinancesApplicationDTOsShowTransactionDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarlonGracieBackendSharedDomainEnumsTransactionType } from './carlonGracieBackendSharedDomainEnumsTransactionType';


export interface CarlonGracieBackendFinancesApplicationDTOsShowTransactionDTO { 
    id?: string;
    type?: CarlonGracieBackendSharedDomainEnumsTransactionType;
    transactionCategoryId?: string;
    transactionCategoryName?: string | null;
    description?: string | null;
    amount?: number;
    transactionDate?: string;
    reference?: string | null;
    createdAt?: string;
}
export namespace CarlonGracieBackendFinancesApplicationDTOsShowTransactionDTO {
}