// src/app/generated_services/model/carlonGracieBackendSharedDomainEnumsTransactionType.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CarlonGracieBackendSharedDomainEnumsTransactionType = 'Debit' | 'Credit' | 'Refund' | 'Adjustment' | 'Income' | 'Expense';

export const CarlonGracieBackendSharedDomainEnumsTransactionType = {
    Debit: 'Debit' as CarlonGracieBackendSharedDomainEnumsTransactionType,
    Credit: 'Credit' as CarlonGracieBackendSharedDomainEnumsTransactionType,
    Refund: 'Refund' as CarlonGracieBackendSharedDomainEnumsTransactionType,
    Adjustment: 'Adjustment' as CarlonGracieBackendSharedDomainEnumsTransactionType,
    Income: 'Income' as CarlonGracieBackendSharedDomainEnumsTransactionType,
    Expense: 'Expense' as CarlonGracieBackendSharedDomainEnumsTransactionType
};